+++ applovin_max/ios/Classes/AppLovinMAXAdView.h
@@ -17,6 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
                        viewId:(int64_t)viewId
                      adUnitId:(NSString *)adUnitId
                      adFormat:(MAAdFormat *)adFormat
+         isAutoRefreshEnabled:(BOOL)isAutoRefreshEnabled
                     placement:(nullable NSString *)placement
                    customData:(nullable NSString *)customData
                     messenger:(id<FlutterBinaryMessenger>)messenger